Rename insert_measure_action -> insert_column_action
